refactor: split up config-context module
