+++ core/src/config/module-template.ts
@@ -11,7 +11,6 @@ import { baseModuleSpecSchema, BaseModuleSpec, ModuleConfig } from "./module"
 import { dedent, deline } from "../util/string"
 import { GardenResource, prepareModuleResource } from "./base"
 import { DOCS_BASE_URL } from "../constants"
-import { ProjectConfigContext, ModuleTemplateConfigContext, EnvironmentConfigContext } from "./config-context"
 import { resolveTemplateStrings } from "../template-string"
 import { validateWithPath } from "./validation"
 import { Garden } from "../garden"
@@ -21,6 +20,8 @@ import { readFile, ensureDir } from "fs-extra"
 import Bluebird from "bluebird"
 import { TemplatedModuleConfig, templatedModuleSpecSchema } from "../plugins/templated"
 import { omit } from "lodash"
+import { ProjectConfigContext, EnvironmentConfigContext } from "./template-contexts/project"
+import { ModuleTemplateConfigContext } from "./template-contexts/module"
 
 const inputTemplatePattern = "${inputs.*}"
 const parentNameTemplate = "${parent.name}"
