+++ core/src/config/workflow.ts
@@ -12,7 +12,7 @@ import { DEFAULT_API_VERSION } from "../constants"
 import { deline, dedent } from "../util/string"
 import { defaultContainerLimits, ServiceLimitSpec } from "../plugins/container/config"
 import { Garden } from "../garden"
-import { WorkflowConfigContext } from "./config-context"
+import { WorkflowConfigContext } from "./template-contexts/workflow"
 import { resolveTemplateStrings } from "../template-string"
 import { validateWithPath } from "./validation"
 import { ConfigurationError } from "../exceptions"
