+++ core/src/tasks/resolve-provider.ts
@@ -19,7 +19,7 @@ import { resolveTemplateStrings } from "../template-string"
 import { ConfigurationError, PluginError } from "../exceptions"
 import { keyBy, omit, flatten, uniq } from "lodash"
 import { GraphResults } from "../task-graph"
-import { ProviderConfigContext } from "../config/config-context"
+import { ProviderConfigContext } from "../config/template-contexts/provider"
 import { ModuleConfig } from "../config/module"
 import { GardenPlugin } from "../types/plugin/plugin"
 import { joi } from "../config/common"
